refactor(policies): extract search highlight extension and toolbar types
Move TipTap search decorations and ToolbarKey defaults into co-located PolicyEditor modules.
